Move ConfirmForm.vue to utils
